Update register_artefact to use new PUT semantics.
We no longer need to check whether an artefact exists before we try to
PUT it: the response code will tell us whether it has been created or
updated. If the current state of the artefact is such that the update
cannot be applied, it will return a 409 response, which will cause an
exception.