Add description as a mandatory field.
As applications port to the new adapters, they will need to include
descriptions. While this is a functional change and as such would
normally require bumping the minor (or even major) version, this is
actually a patch as the behaviour in the 0.1.0 release was incorrect.

We'll probably need to be more careful with this once we hit 1.0.0.