Break up GdsApi::JsonClient#do_request a bit
This was an overly complex method. I've not improved it significantly, but I've moved
a lot of the conditional behaviour out to their own methods so hopefully the core
code reads a little more easily.

I'm mainly doing this as I'm about to add another auth option that would have tangled
the code even further.