Update LicenceApplication API to conform with Licensify
The original work done to produce the initial LicenceApplication API was
put in place to provide an example of the methods that would be used in
other applications.

This commit updates the core implementation of the API, adding an extra
method to expose features of the API that have since been added. This
change is backwards compatible, but should invoke a minor version
increase when it gets exposed to the gem.