Specify the endpoint url directly instead of the platform
This is part of the Great Plek Purge. GdsApi adapters should require the endpoint url
to be passed in, and shouldn't construct endpoint urls based on the platform. This
is a substantial API change and so will require a version bump. It is intended that
callers of this module can specify an endpoint url using the Plek with
Plek.current.find(appname) -- which will soon be replaced with Plek.new.find(appname)