Make Artefact stubbing permissive on query params
Previously, requests had to have exactly the specified query params,
now they just have to have the params you specify.

Specific motivation is that lots of requests in frontend will soon
include latitude and longitude, but it's not something that matters
in most of the tests, so it's nicer if the stub doesn't care.