Ensure tests get consistent Plek-generated URLs
This ensures that the tests pass in CI and dev with Plek < 1.0.0, which
they haven't. With Plek 1.0.0 the problem should go away, but this code
should probably stay for backwards compatibility.

If you want more detail on what the problem was:

CI would get contentapi.production... and dev would get contentapi.test...
this meant that tests dependent on URL length would fail in CI, even
though they passed in dev.

Plek < 1.0.0 uses the RAILS_ENV/RACK_ENV to determine if it should
return test URLs. If it's not set to "test", it uses FACTER_govuk_platform
then RAILS_ENV/RACK_ENV, then "development".
This app doesn't use RAILS_ENV/RACK_ENV and because FACTER_govuk_platform
is set to "production" on CI, but not set in dev, different URLs were used
in each environment.