Add post_multipart method to JsonClient
This changes the existing behaviour of do_raw_request
to return the entire http response, and not just the
body. 

This makes get_raw more consistent with get_*/post_*
methods in JsonClient in returning a response. 
(However, get_raw does not return a GdsApi::Response 
due to the assumptions it makes around parsing JSON.)

From a search of the code, the only method using this
is the Imminence API's places_kml method, which I've
updated in the preceding commit.