Clearer naming for imminence test helper methods
The original names imply that the methods are setting up some kind of
registry of available data, whereas they are actually simply setting up
data that will be returned for an exact match on a query.

This should reduce the risk of further confusion.

This is a change to the published api so will require a major version
bump when merged as per http://semver.org/