Decouple tag methods from the section tag type
In the Content API adapter, extract out the methods for making requests 
about tags, so that they are no longer tied to the section tag type and
accept an argument of the tag type to look up.

For clients which use the 'sections', 'root_sections' and 'sub_sections'
methods, the existing behaviour will continue to work.