Decouple tag test helpers from the section tag type
Create new test helpers for tags in the Content API which can be used 
with any tag type, not just the section tag.

The existing test helper methods will continue to work, however there is
a breaking change to existing tests as the ID for a tag will now always
include the tag type (previously it used the legacy behaviour where a 
tag lived at '/tags/foo.json').

There's a hacky bit of pluralisation here to deal with the way that the
Content API pluralises the tag type in URLs. Ideally this can go away
soon.