Prevent branch builds publishing the gem.
To publish the gem after a successful build, we now need to set the
`PUBLISH_GEM` environment variable.

This changes the conventional behaviour of `jenkins.sh`, but as @alext
pointed out, it's better to break convention and fail safely than to
keep with convention and fail dangerously.

Branch builds technically shouldn't end up publishing anyway, so long as
everyone follows the rule of not bumping the gem version on a branch
build, but it's better not to take the risk.