Router client no longer commits routes by default.
This reverses the sense of the option parameter to allow optionally
requesting to commit as well.  This is a breaking change, and will be
released as a major version bump.

The reason for this is that everywhere we use this at the moment, we're
registering more than one route, and therefore using the skip_commit
option.  It therefore makes sense to make this behaviour the default.