Require exception classes
We were referencing the GdsApi::BaseError class in our code, without explicitly
requirin 'gds_api/exceptions'. In development mode this wasn't a problem, because
an adapter class, which does require it, was always loaded first. When we deployed
to preview however, the application couldn't boot because the constant was
undefined.