Add new Panopticon test helpers (with tests!)
This adds three new test helpers to stub out the requests that clients
will make to Panopticon to create and update tags.

I've added unit tests for the test helpers here, which is something that 
hasn't previously been done before in this library. The tests assert 
that the correct stubs have been registered with WebMock for each test
helper.