Pass correct parameters to assert_requested
Previously, this was just passing an options hash to assert_requested,
which isn't correct usage, and would have been failing to actually check
the body.

Webmock now (as of release 1.20.3) validates the options passed to
assert_requested, and will complain if unknown keys are passed.  This
has made the error obvious.