Support pagination in Collections API test helper
This changes the Collections API test helper to support pagination. 

It includes the 'documents_start' and 'documents_total' keys in the 
response body, so that apps using the test helper can test their 
pagination behaviour.

It also allows each of these values to be overridden by providing the
'return_body_options' key in the 'options' argument. This is so that 
tests can simulate scenarios with multiple pages of results.