Add a test helper for Rummager-with-queueing
Rummager returns 202 on POST if a feature flag is enabled, because it
queues the change:

https://github.com/alphagov/rummager/blob/cc0e9c95db9098b0dbc3dd12124d1eef5a153ae7/app.rb#L460-L466

This queueing is disabled in development, so this request returns a
different status code from in production. We need to be able to test
that other applications interact with Rummager as expected in all
environments, so this commit adds a test helper to simulate Rummager's
behaviour when queueing is enabled.