Adapter for sending drafts to publishing API
https://trello.com/c/PbCzGBIB

Content designers need a way to preview content they're drafting.
When draft content is saved, publishing applications can push it
to a draft instance of the content store, from where it can be
previewed on the draft site.

To begin with, this call will return an empty 200 OK response,
unless there's an instance of draft content store configured in
the environment and Publishing API is told about it by setting
DRAFT_CONTENT_STORE in the environment.

See related Publishing API change:
https://github.com/alphagov/publishing-api/pull/22