Fix content_api_has_artefacts_with_a_draft_tag helper.
It was broken with the changes to content_api_has_artefacts_with_a_tag
which was calling artefact_for_slug with nil instead of empty hash
causing things to blow up.

content_api_has_artefacts_with_a_draft_tag was also using the now
deprecated method of specifying that the tag is draft.