Add pact setup code and provider definition
The PactTest mixin defines any general setup/teardown steps needed
for pact contract tests. We want to disable webmock during pact tests as
it will interfere with pacts own mocking of HTTP.

The provider definition specifies the `publishing_api` provider which
runs on port 3093. This is part of the mechanism for stubbing the
service and recording the interaction with it.