Remove support for deprecated contentapi path
This has been deprecated since:
https://github.com/alphagov/govuk_content_api/commit/5361729cd9adb4fe8ba4d94d6e428b6fc8fc8465

Using grep and looking at the user agent in the logs for contentapi in
production, the only place I was able to find this being used was in Whitehall
in the code for building breadcrumbs. We should be able to switch this to
specify the tag_type as it will always be looking for tags of type "section".

This will require a MAJOR version change as it changes the API in backwards
incompatible way.