Raise specific error on 404 in content-store client.
In order to allow this to be rescued in controllers without also
rescuing 404s from other API clients. Because content-store is the
primary source of information about what is available at a given path,
it makes sense to pass on 404s from content-store, whereas it may not
for other API clients.