Fix timezone handling
if Time.zone returned nil, which unexpectedly does happen in some
circumstances, then the registration would crash.

This makes explicit the expected behaviour.

I've also changed the tests to use Timecop to freeze the system clock to
avoid possible intermit false failures in the tests (due to the clock
happening to roll over the 1 second boundary between the setting of the
expectation and the publishing of the special route).