Remove Rummager#organisations
This removes the Rummager endpoint for organisations.

The `/organisations` endpoint was used by the frontend application to
render the select box of organisations. It was added in
https://github.com/alphagov/rummager/pull/103 (May 2013).

Frontend stopped using the endpoint since moving to "Unified Search",
and the call to rummager was removed in PR
https://github.com/alphagov/frontend/pull/601.

PR: https://github.com/alphagov/rummager/pull/494