Add `HTTPConflict` exception
HTTP 409 can be used by client to indicate that the issued request is
conflicting with something else.

To be used by panopticon tell the user that a tag can't be deleted
because it still has documents tagged to it.

> The request could not be completed due to a conflict with the current
state of the resource. This code is only allowed in situations where it
is expected that the user might be able to resolve the conflict and
resubmit the request.