Set both kinds of timeout in RestClient
Setting `timeout` only sets the read timeout, which is the period
between a request being issued and receiving the response. We’ve had
issues with licensify in staging and new production where connecting to
the remote host waits for over a minute before erroring, so setting the
connection timeout will allow the request to fail faster.

The default timeout is 4 seconds, which seems reasonable for the
initial SYN/ACK cycle for a remote host.

In version 2.0.0 of RestClient (currently unreleased), both of these
timeouts will be set via the `timeout` variable, but can also be
individually set with `read_timeout` and `open_timeout`. Therefore,
setting `timeout` and `open_timeout` here is forward-compatible. The
change can be reverted once we upgrade to RestClient 2.0.0.