Return 200 not 201 from publishing api stubs
The publishing API delete commands never returned a 201 status code,
they returned a 200 status code.

The PUT /content and PUT /draft-content calls used to return 201 for a
new item and 200 for an existing item, but that behaviour changed when
we switched to a rails implementation of publishing api. This feature
was not tested anywhere so it wasn't noticed at the time. We don't
believe any client applications relied upon this behaviour so we've
encoded this expectation in the new contract tests[1]. This commit
updates the stub helpers to reflect the actual behaviour.

[1]
https://github.com/alphagov/gds-api-adapters/commit/2f74464a838f51a0d4c182e18b5ebd6f88242826#diff-a8f8b79ce21998f08dbaa06c40cc4551R31