Publish pacts to the pact broker.
This adds hooks to publish the generated pacts to the pact-broker after
the test run.

If PACT_TARGET_BRANCH is specified, it'll push the pacts to that target.
This will be used to push to branch targets (master, or branch-foo) so
that this project can then trigger builds of other projects using these
pacts before marking the build as green.

This additionally adds a hook to publish the pacts corresponding to the
released version when the gem is published.