Mapit raises error on 404
Previously, the Mapit adapter was only raising the error when the API
return a 400. This commit updates it to also return on a 404, which
happens when a User enters a valid postcode but gets back no matches.
Most commonly, this happens with new build housing developments which
haven't been added to the MapIt DB.
[Ticket](https://trello.com/c/Db6csIA7/135-log-the-errors-users-see-when-entering-postcodes-on-local-transaction-pages).