Remove Hash deep_stringify_keys patch
The Hash patch didn't support `Array` and was also breaking other things
once the gem got into a Rails app as it was conflicting with the rails
implementation.

The working version has been moved into the test helper file where it
was needed.

This can be folded into f97c6c7 before merge.