Include the request in the error message
This should make the default error reporting much more useful. At the moment,
it can be very difficult to figure out what the request was that triggered the
error. Whilst clients could explicitly capture and record the problem request
body, that requires lots of manual work in every place a gds-api-adapter is
used.

The downside is that the error message suddenly becomes much longer. I can't
really see how that is a problem.