Switch PublishingApiV2 put_links to patch_links
After speaking to Daniel Roseman and Richard Boulton, we agreed that this
endpoint's behaviour and the original intent was for it to be a PATCH action,
not an overwrite (PUT).

By naming it to better reflect the behaviour, we should avoid confusion.

The publishing-api will support PUT until at least all of the clients have
moved to PATCH, possibly forever.