Don't test for request body if the argument is not supplied
The default empty hash argument for `attributes_or_matcher` means
that the assertion only passes when the client has sent an empty hash to
the publishing-api.

In most cases where the second argument is not supplied, we don't care
about the payload. Webmock will not test the request body if it is nil.

To summarise, after this change:

`assert_publishing_api_put_links(content_id)` will pass with any body
`assert_publishing_api_put_links(content_id, {})` will only pass if the
body is empty