Remove pact for an unreachable state…
The Publishing API validates that publishing_app
must be present on the content item, so this state
isn’t reachable (unless validation is bypassed in
the provider states).

I don’t think we should be testing how the system
behaves when it’s been placed into an inconsistent
state. I’d rather it was unspecified, if this ever
does occur, rather than having to support some
weird edge case.