Remove public_updated_at from assertions…
Previously, this field was always returned because
it happens to be the default sort order, but the
queries in publishing-api have been reworked such
that they can order by a field that doesn’t have
to be included in the output.

Therefore, the interface has changed slightly and
fields must be explicitly requested rather than
being included by default.