Update links to rubydoc.info
The current links give a little unpredictable results (you often see the docs for 30.X, while were at 30.2.X). I've set up a GitHub webhook from this repo, so master should always be up to date. It also allows easy switching to the docs for a branch, which could be nice in code review.