Remove implicit dependency between JsonClient & GdsApi::Base
Previously `JsonClient` used `GdsApi::Base.logger` as its default logger.
However, there was no corresponding `require 'base'` statement at the top of the
file. This caused a problem [1] in Smart Answers which uses `JsonClient`
directly.

It's not possible to fix this by adding such a `require` statement, because this
would introduce a circular dependency. So as an alternative, I've used
`NullLogger.instance` as the default value, but at the same time ensured that
`GdsApi::Base` always specifies a logger in the options passed to
`JsonClient#initialize`. I'm pretty confident this means that the behaviour
hasn't really changed and yet there is no longer a dependency between
`JsonClient` and `GdsApi::Base`.

[1]: #2538