Always raise when publishing-api & content-store return not found
This commit makes sure that the publishing-api and content-store
clients rais a `GdsApi::NotFound` error when the server returns a 404
(Not Found) or 410 (Gone).

This was previously only the case for some methods. In some cases a
pattern was to have a `bang!` method and a non-`bang!` method, so that
the client could opt-in to this behaviour. Clients like Rummager use
the `get_json!` method exclusively.

Reasons to change:

- This makes the behaviour consistent across the most used clients
(publishing-api, content-store and rummager)
- Less duplication between `bang!` and `non_bang` methods.
- Makes it less likely that we inadvertently introduce nils, which
generate errors that are hard to spot and to debug.