Change Publishing API's port in tests
We were using port `3093` in the tests for a mocked version of the Publishing
API. That is also the port used in development mode. This means that we run the
publishing api in development mode and run the gds-api-adapters' tests, we get a
bunch of failures, as the code hits the actual Publishing API endpoints.

This commit changes the port of the test Publishing API and makes it easier to
change it later on by using a method to return the localhost URL of the
publishing API.