Add option to resturn results as a hash
This change checks if GdsApi.config.hash_response_for_requests is set to true, and returns a results hash instead of an OpenStruct in ListResponse.  We discovered this was necessary after opting in to the new GDS API Adapters behaviour in contacts-admin (see PR [here](https://github.com/alphagov/contacts-admin/pull/248) and Trello card [here](https://trello.com/c/B5z6Dq8q/495-get-our-apps-up-to-date-on-new-api-adapters-behaviour-update-at-stand-up), and found that tests were failing due to results returning an OpenStruct.