Only return expected number of items from publishing_api_has_content
This change makes sure that we return the expected results based on the
pagination information.

Even though we need to stub the API call with a number of items, this endpoint
should only return the ones belonging to the page requested (taking into account
the number of items per page).

As an example, for a call like:

```
publishing_api_has_content(
  [
    { "content_id" => 'content_id_1' },
    { "content_id" => 'content_id_2' },
    { "content_id" => 'content_id_3' },
  ],
  {
    page: 2,
    per_page: 2
  }
)
```

Where the user is requesting the 2nd page with 2 items per page, the API call
should only return the last entry in that array ('content_id_3'), which is the
only items on the 2nd page.

For out-of-bound queries, we should return an empty array.