Default always_raise_for_not_found to true
We are in the process of deprecating `GdsApi.config.always_raise_for_not_found`.
Initially, we added a deprecated message encouraging all clients to set that
config option to `true`. By doing that, we will raise exceptions for 404s and
410s.

This change now makes that option `true` by default and moves the deprecation
message to the setter of `always_raise_for_not_found`.

By December 1st, 2016 we will be removing that configuration option and make
sure we always raise for 404s and 410s.