Default hash_response_for_requests to true
We are in the process of deprecating `GdsApi.config.hash_response_for_requests`.
Initially, we added a deprecated message encouraging all clients to set that
config option to `true`. By doing that, we will return a response that behaves
like a Hash instead of an OpenStruct.

This change now makes that option `true` by default and moves the deprecation
message to the setter of `hash_response_for_requests`.

By December 1st, 2016 we will be removing that configuration option and make
sure we always return an object that behaves like a Hash instead of an
OpenStruct.