Use "manual" rather than "special_route"
The tests relating to previous_version do not work for "special_route"
now that schema validation is taking place, as the schema for
"special_route" does not allow previous_version in the payload.

Therefore, switch to using "manual" to test this functionality as this
allows previous_version, and looked like it didn't place many
requirements on details.