Generate docs only for documented APIs
Most of the classes in this application aren't documented. This means
that the documentation isn't really useful for the most part. This
commit adds a `.yardopts` file that is used by YARD to configure the
doc build. The only classes it will document from now on are classes
that have at least some documentation.

My hope is that by making the docs smaller they'll be more useful and
in turn convinces people to add more docs.