Fix linting errors
In https://github.com/alphagov/gds-api-adapters/pull/628 I'm attempting
a large scale refactor. The linting won't pass for that PR because
there are styleguide violations in all the files it's touching (we
normally only run rubocop against changed files).

This fixes all styleguide violations so that that PR can be merged
without noise.