Publish the released version of the pact.
This used to be part of the publish_gem rake task, but that job has
been failing since the move to Jenkins 2 - although the gem is
successfully published, it reports an error and fails to proceed, so
never invokes the pact:publish:released_version task.

This change moves the task out into a separate Jenkins stage, which
makes it clearer what is happening anyway.