Allow publishing_api_has_expanded_links to take content_id as string
The test helper for publishing_api_v2, publishing_api_has_expanded_links,
takes a hash as its argument. The content_id key inside this hash currently
needs to be a symbol. This commit allows the content_id to be either a
symbol or a string. This makes the method consistent with other helper
methods and also with the docs, which show content_id as a string.