Check out content schemas before all tests are run, not just pub-api
Before, govuk-content-schemas was checked out before the publishing-api
Pact tests were run. This checkout has been moved to the start of the CI
build so that the gds-api-adapters can also use the content schemas when
running tests.

This is in preparation for a later change which will add some tests that
depend on the content schemas.