Update webmock dependency
Fix tests which are broken by the upgrade. In version 1.24.3, Webmock
became stricter about matching the body of `multipart/form-data`
requests because it could not match the body in all cases even though it
worked for some cases like the tests in this project.

The tests have been fixed by using Webmock's more general request
matching which takes a block.