Make LGIL code mandatory for local links requests
So that we can make administering links easier, we want to be able to only
support a defined set of LGSL / LGIL combinations.  We currently have some
logic to find the "best" result if an LGIL code isn't supplied with a request
for a link, but this results in some tricky code, more links to support and
uncertainty about whether the links that we provide are as good as they might
be.

To do this we need to:

- Assign LGIL codes to all `local_transaction` formats in
Publisher (done)
- Make this change to GDS API adapters (and bump the major version)
- Bump the Gem in Frontend and make sure that everything works
- Make the LGIL code mandatory on the LLM api