Rename support_api test helpers to reference support_api
Some of the test helpers had been copied from the support test helper
and so looked like they were stubbing requests to the support app (for
example: stub_support_long_form_anonymous_contact_creation).  Some had
no indication of which app they stubbed requests to (for example:
stub_create_page_improvement).  By including suppor_api in the name of
the helper we make it clearer which app they are stubbing which is
useful in test scenarios where we might include several api test helpers.
We also achieve consistency with other adapters which include the
name of the api in their helpers.