Add an API endpoint for info-frontend
For: https://trello.com/c/9JNitfXS/173-spike-info-pages

This will add an api endpoint for info-frontend to make it possible for it
to talk to the Backdrop Read API.

This will enable us to bypass metadata-api for making these api calls, and
will ultimately allow us to retire metadata-api, since it it currently only
used for these api calls.

This commit will be followed with a commit in info-frontend to use these
endpoints.