Add `get_paged_editions` method
This change adds a method for paging through responses from the
`/v2/editions` endpoint.

The `editions` endpoint is paged through its returned `next` link,
which doesn’t work well with the `gds-api-adapters` pattern. This
change rolls the pagination into a method that returns a lazy
`Enumerator` that knows how to fetch the next page.