Resurrect `GdsApi::Support#feedback_url`
We removed this in 1a3694ff3f797dc1906ce598d5228a7e5c3fd20e as we thought
nothing used it and it referred to a path the support app didn't respond
to.  Both of these facts turned out to be false; maslow uses it to link
to the feedback for pages that are covered by a need, and the app does
respond to that url.