Avoid adding Content-Type header for multipart requests
If `params[:multipart]` is `true` then `RestClient::Payload.generate`
creates a `MultiPart` payload which sets the `Content-Type` header to
`multipart/form-data`. This would override the `Content-Type`
header from `default_request_with_json_body_headers` and result in a
warning like:

  warning: Overriding "Content-Type" header "application/json" with
  "multipart/form-data; boundary=----RubyFormBoundaryX7Na6WDQqG3kLfD7" due
  to payload

I have to check whether `params` is a hash-like object because it can
also be a string in the case that `#do_request` is called from
`#do_json_request`.

I'm not particularly keen on the test I've added but I couldn't really
see how else to do it. I considered testing for the absence of a warning
but that doesn't feel quite right.