Move JSON-specific header to do_json_request
My previous commit increased the complexity of the code around
`default_headers` in the `#do_request` method. This commit attempts to
reduce that complexity.

We only want to set the `Content-Type` to `application/json` if we're
sending JSON in the body of our request. We were previously doing this
in the `#do_request` method for non-GET and non-DELETE requests (i.e.
for POST, PUT or PATCH requests) where that request wasn't explicitly
being set to multipart. I think it makes the intention clearer, and the
code simpler, by setting this 'Content-Type' header in the
`do_json_request` method instead.

I'm setting the `Content-Type` header in `do_json_request` based on the
presence of the `params` argument alone (i.e. I'm ignoring the
`method`). This matches the previous behaviour (described above) because
`params` is only set for POST, PUT and DELETE requests (i.e. we won't
set the header for GET or DELETE requests).