Document optional legacy_* params for create_whitehall_asset
In alphagov/asset-manager#242, the Asset Manager API was changed to accept
these optional attributes. The idea is that they will only be used to
move *existing* Whitehall assets into Asset Manager in order to avoid
wholesale cache invalidation. They should not be used when creating
*new* Whitehall assets.

Note that I didn't need to change any of the code to support this,
because the `asset` Hash parameter already allowed arbitrary attributes
to be passed into the `GdsApi::AssetManager#create_whitehall_asset`
method.