Lock down webrick to v1.3.1
This will enable us to have a succesful build when testing the gem against ruby versions 2.1 and 2.2.

Long explanation: A new version of webrick has been launched (1.4.0) which gets installed by `pact-consumer-minitest` (it installs the latest version it can find).
Unfortunately webrick 1.4.0 and up will only work with ruby >= 2.30, but we also test for ruby versions 2.1 and 2.2 in our Jenkinsfile.

In order to get around this constraint, we've had to lock down webrick in the .gemspec file to version 1.3.1 for now.

Our next step will probably involve removing ruby versions 2.1 and 2.2 from the Jenkinsfile, removing this webrick lockdown and adding potentially adding ruby 2.4 to test against.