Change disable cache env var name
Since this env var gets set in applications it follows a precedent to
prefix this with GDS_API to indicate it's for a specific library.

The rest of the environment variable has been made more generic as there
is only one kind of cache and the anticipation is that if you needed
this env var you'd want to disable this and future types of cache.