Helper method to determine destination of redirect
This is a helper for ContentStore adapter. It allows resolving, based on
a request, where a redirect is meant to resolve to.

This logic is performed in the router under normal circumstances,
however the transition of content to a redirect is not atomic and it is
possible (and has happened) that requests meant for router will be
forwarded to frontend apps.

The intention of this code is to provide a helper method that can give a
frontend app the opportunity to do the redirect themselves if they wish
with minimal understanding of the logic behind it.